+++ src/game/Coordinate.java
@@ -3,27 +3,27 @@ package game;
 
 public class Coordinate {
 
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public Coordinate(int x, int y) {
+    public Coordinate(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
